Quick Interface ...
...is a child class of cgi app that should provide an extreemly
simplified interface to the QTC Network as part of the webapp
so applications do not have to implement the whole QTC net protocol.

It will get parameters via POST Request, and returns an XML File
with all the relevant data that can be parsed by the application.

STATE: initial commit not yet working